feat : creation des entites et repositories

+++ src/main/java/com/application/shoppinglistmanager/ingredients/IngredientsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngredientsRepository extends JpaRepository<Ingredients, Long>{
+public interface IngredientsRepository extends JpaRepository<Ingredients, Integer>{
     
 }
